+++ pkg/cortex/cortex.go
@@ -365,7 +365,6 @@ func New(cfg Config) (*Cortex, error) {
 
 	// Swap out the default resolver to support multiple tenant IDs separated by a '|'
 	if cfg.TenantFederation.Enabled {
-		util_log.WarnExperimentalUse("tenant-federation")
 		users.WithDefaultResolver(users.NewMultiResolver())
 	}
 
@@ -524,7 +523,7 @@ func (t *Cortex) readyHandler(sm *services.Manager) http.HandlerFunc {
 
 			byState := sm.ServicesByState()
 			for st, ls := range byState {
-				msg.WriteString(fmt.Sprintf("%v: %d\n", st, len(ls)))
+				fmt.Fprintf(&msg, "%v: %d\n", st, len(ls))
 			}
 
 			http.Error(w, msg.String(), http.StatusServiceUnavailable)